+++ cate/util/web/webapi.py
@@ -251,7 +251,7 @@ class WebAPI:
 
         print(f'{name}: started service, listening on {join_address_and_port(address, port)}')
 
-        self.server = application.listen(port, address=address or 'localhost')
+        self.server = application.listen(port, address='' if address == 'localhost' else address)
         # Ensure we have the same event loop in all threads
         asyncio.set_event_loop_policy(_GlobalEventLoopPolicy(asyncio.get_event_loop()))
         # Register handlers for common termination signals