+++ CHANGES.md
@@ -1,4 +1,9 @@
-## Version 2.0.0.dev11 (in development)
+## Version 2.0.0.dev12 (in development)
+
+* The operation  `long_term_average` now works with daily, monthly and seasonal datasets [#471](https://github.com/CCI-Tools/cate/issues/471)
+* Fixed problem in `cate-webapi-start` occurring on Linux when using address `localhost` (related to [#627](https://github.com/CCI-Tools/cate/issues/627)) 
+
+## Version 2.0.0.dev11
 
 * Lacking cancelable progress monitor when opening large datasets [#640](https://github.com/CCI-Tools/cate/issues/640)
 * Wrong chunk size does not allow to import some dataset [#631](https://github.com/CCI-Tools/cate/issues/631)